rebuild the oneclick installer on each merge to main
We will be building the Oneclick installer with Juno-Boostrap included to facilitate airgapped installs.
That means we need to keep things up to date automatically - each merge to main on Juno-Bootstrap should also trigger a release of the oneclick installer.

I will open a separate PR in K8s-Playbooks to listen for/accept the dispatch event we
define here.