release the install helper as a self-contained package, refactor a bit
This packages the install helper as a single bash file, removing the
need for online dependencies.

I also have done a few small refactors and twaeked some minor handling
that jumped out at me when playing with it.
(eg. re-prompt instead of quitting on invalid selection)

This PR only saves a workable chunk of progress. The current (online)
functionality works, for offline we'll need to make adjustments in the
charts themselves - that doesn't work quite yet!

The registry mirror handling for offlien installs is already fully
working and verified on my airgap lab setup. Only the charts remain.