bump gpu-operator to 25.10.1 to support ubuntu24
Currently ubuntu24 will fail to install drivers.
For customers that wish to use it, we must bump it.

I already did some practical validation on this.
We'll also need to make sure the OneClick sets the containerd path for
k3s.

I'd rather force that in the juno_k3s role instead of here. I'll bump
the change there.